requirements.txt
gradio
openai
python-dotenv
pillow
numpy
beautifulsoup4
feedparser
google-generativeai
pandas 
openpyxl
xlrd
pymupdf
python-docx
